Fix uint64 casts on FreeBSD

+++ utils/storage/storage_unix.go
@@ -15,6 +15,6 @@ func AvailableBytes(storagePath string) (uint64, error) {
 		return 0, err
 	}
 	//	avail := stat.Bavail * uint64(stat.Bsize)
-	avail := uint64(stat.Bavail * int64(stat.Bsize))
+	avail := uint64(stat.Bavail) * stat.Bsize
 	return avail, nil
 }
